% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roughsf.R
\name{roughsf}
\alias{roughsf}
\title{Create a rough map}
\usage{
roughsf(
  object,
  title = NULL,
  roughness = 1,
  bowing = 1,
  simplification = 1,
  font = "30px Arial",
  width = NULL,
  height = NULL,
  elementId = NULL,
  chunk_name = "canvas"
)
}
\arguments{
\item{object}{sf object}

\item{title}{optional title of the map}

\item{roughness}{numeric vector for roughness of lines}

\item{bowing}{numeric vector for bowing of lines}

\item{simplification}{simplify drawings (remove points from objects)}

\item{font}{font size and font family for labels}

\item{width}{width}

\item{height}{height}

\item{elementId}{DOM id}

\item{chunk_name}{markdown specific}
}
\description{
plot a sf map using rough.js
}
\details{
the function recognizes the following attributes:
\itemize{
\item \emph{fill} shape fill color
\item \emph{color} shape stroke color
\item \emph{stroke} stroke size
\item \emph{fillstyle} one of "hachure", "solid", "zigzag", "cross-hatch", "dots", "dashed", "zigzag-line"
\item \emph{label} label (not implemented yet)
}

Default values are used if one of the attributes is not found.

The result of a roughnet call can be printed to file with \code{save_roughnet()}

More details on roughjs can be found on https://github.com/rough-stuff/rough/wiki
}
