% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_roughsf.R
\name{save_roughsf}
\alias{save_roughsf}
\title{Save roughsf plot to file}
\usage{
save_roughsf(rsf, file, background = "white")
}
\arguments{
\item{rsf}{result from calling the function \code{roughsf}}

\item{file}{filename}

\item{background}{string giving the html background color}
}
\description{
Save roughsf plot to file
}
